+++ build-logic/convention/src/main/kotlin/AndroidLibraryConventionPlugin.kt
@@ -1,0 +1,32 @@
+import com.android.build.gradle.LibraryExtension
+import com.smileidentity.buildlogic.configureKotlinAndroid
+import com.smileidentity.buildlogic.libs
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.apply
+import org.gradle.kotlin.dsl.configure
+import org.gradle.kotlin.dsl.dependencies
+
+class AndroidLibraryConventionPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        with(target) {
+            apply(plugin = "com.android.library")
+            apply(plugin = "org.jetbrains.kotlin.android")
+            apply(plugin = "org.jlleitschuh.gradle.ktlint")
+
+            extensions.configure<LibraryExtension> {
+                configureKotlinAndroid(this)
+                defaultConfig.targetSdk = 36
+                defaultConfig.testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+                testOptions.animationsDisabled = true
+            }
+
+            dependencies {
+                "implementation"(libs.findLibrary("timber").get())
+                "androidTestImplementation"(libs.findLibrary("androidx.test.junit").get())
+                "androidTestImplementation"(libs.findLibrary("kotlin.test").get())
+                "testImplementation"(libs.findLibrary("kotlin.test").get())
+            }
+        }
+    }
+}