+++ sample/src/main/java/com/smileidentity/sample/compose/components/IdTypeSelectorAndFieldInputScreen.kt
@@ -57,7 +57,7 @@ fun IdTypeSelectorAndFieldInputScreen(
     viewModel: IdTypeSelectorAndFieldInputViewModel = viewModel(
         factory = viewModelFactory { IdTypeSelectorAndFieldInputViewModel(jobType) },
     ),
-    onResult: (IdInfo, ConsentInformation) -> Unit,
+    onResult: (IdInfo, ConsentInformation?) -> Unit,
 ) {
     val uiState = viewModel.uiState.collectAsStateWithLifecycle().value
     Box(