+++ ui/src/main/java/com/smileidentity/ui/components/FaceShapedView.kt
@@ -1,0 +1,8 @@
+package com.smileidentity.ui.components
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun FaceShapedView(modifier: Modifier = Modifier) {
+}