+++ CHANGELOG.md
@@ -5,10 +5,16 @@
 ### Added
 
 * Added Face Tracking in Selfie Capture to ensure only same face is used during capture
+* Added `autoCaptureTimeout` that allows partners to override the default 10 seconds auto capture
+  timeout
 
 ### Changed
 
 * Removed `AntiFraud` response in `JobStatus` calls
+* Updated the `targetSdk` to 36 and updated the AGP version
+* Removed the default `ConsentInformation` set when no consent is passed
+* Changed `enableAutoCapture` to `AutoCapture` enum to allow to allow partners change document
+  capture options
 
 ## 11.0.5 - July 2, 2025